refine the code test=develop
