+++ paddle/fluid/operators/gaussian_random_op.cc
@@ -38,7 +38,8 @@ class CPUGaussianRandomKernel : public framework::OpKernel<T> {
     engine.seed(seed);
     std::normal_distribution<T> dist(mean, std);
 
-    auto shape = GetShape(context);
+    const std::string op_type = "gaussian_random";
+    auto shape = GetShape(context, op_type);
     tensor->Resize(shape);
     int64_t size = tensor->numel();
     T* data = tensor->mutable_data<T>(context.GetPlace());
