+++ paddle/fluid/operators/gaussian_random_op.cu
@@ -51,7 +51,8 @@ class GPUGaussianRandomKernel : public framework::OpKernel<T> {
     T std = static_cast<T>(context.Attr<float>("std"));
     thrust::counting_iterator<unsigned int> index_sequence_begin(0);
 
-    auto shape = GetShape(context);
+    const std::string op_type = "gaussian_random";
+    auto shape = GetShape(context, op_type);
     tensor->Resize(shape);
     int64_t size = tensor->numel();
     T* data = tensor->mutable_data<T>(context.GetPlace());
