+++ python/paddle/fluid/layers/nn.py
@@ -9620,9 +9620,8 @@ def gaussian_random(shape, mean=0.0, std=1.0, seed=0, dtype='float32'):
                shape2 = fluid.layers.fill_constant(shape=[1], dtype='int32', value=4)
                x1 = fluid.layers.gaussian_random((2, 4), mean=2., dtype="float32", seed=10)
                x2 = fluid.layers.gaussian_random([shape1, 4], mean=2., dtype="float32", seed=10)
-               x3 = fluid.layers.gaussian_random([shape1, shape2], mean=2., dtype="float32", seed=10)
                x1_np = x1.numpy()       
-           x_np
+           x1_np
            # array([[2.3060477 , 2.676496  , 3.9911983 , 0.9990833 ],
            #        [2.8675377 , 2.2279181 , 0.79029655, 2.8447366 ]], dtype=float32)
     """
@@ -9641,50 +9640,9 @@ def gaussian_random(shape, mean=0.0, std=1.0, seed=0, dtype='float32'):
         'dtype': c_dtype,
         'use_mkldnn': False
     }
-    op_type = "gaussian_random"
-    inputs = {}
 
-    def _get_attr_shape(list_shape):
-        attr_shape = []
-        for idx, dim in enumerate(list_shape):
-            if isinstance(dim, Variable):
-                attr_shape.append(-1)
-            else:
-                attr_shape.append(dim)
-        return attr_shape
-
-    def _get_shape_tensor(list_shape):
-        new_shape_tensor = []
-        for idx, dim in enumerate(list_shape):
-            if isinstance(dim, Variable):
-                dim.stop_gradient = True
-                check_dtype(
-                    dim.dtype, 'shape[' + str(idx) + ']', ['int32', 'int64'],
-                    op_type,
-                    '(When type of shape in' + op_type + 'is list or tuple.)')
-                if convert_dtype(dim.dtype) == 'int64':
-                    dim = cast(x=dim, dtype='int32')
-                new_shape_tensor.append(dim)
-            else:
-                temp_out = helper.create_variable_for_type_inference('int32')
-                fill_constant([1], 'int32', dim, force_cpu=True, out=temp_out)
-                new_shape_tensor.append(temp_out)
-        return new_shape_tensor
-
-    if isinstance(shape, Variable):
-        shape.stop_gradient = True
-        check_dtype(shape.dtype, 'shape', ['int32', 'int64'], 'fill_constant',
-                    '(When type of shape in' + op_type + ' is Variable.)')
-        if (convert_dtype(shape.dtype) == 'int64'):
-            shape = cast(shape, 'int32')
-        inputs["ShapeTensor"] = shape
-    elif isinstance(shape, (list, tuple)):
-        assert len(shape) > 0, (
-            "The size of 'shape' in" + op_type + " can't be zero, "
-            "but received %s." % len(shape))
-        attrs["shape"] = _get_attr_shape(shape)
-        if utils._contain_var(shape):
-            inputs['ShapeTensorList'] = _get_shape_tensor(shape)
+    inputs = utils._get_shape_tensor_inputs(
+        helper, attrs=attrs, shape=shape, op_type='gaussian_random')
 
     helper.append_op(
         type='gaussian_random',
